+++ Source/Scene/CustomSensorVolume.js
@@ -120,6 +120,19 @@ define([
         this.modelMatrix = t.modelMatrix || Matrix4.IDENTITY.clone();
 
         /**
+         * <p>
+         * Determines if the sensor is affected by lighting, i.e., if the sensor is bright on the
+         * day side of the globe, and dark on the night side.  When <code>true</code>, the sensor
+         * is affected by lighting; when <code>false</code>, the sensor is uniformly shaded regardless
+         * of the sun position.
+         * </p>
+         * <p>
+         * The default is <code>true</code>.
+         * </p>
+         */
+        this.affectedByLighting = this._affectedByLighting = (typeof t.affectedByLighting !== 'undefined') ? t.affectedByLighting : true;
+
+        /**
          * DOC_TBA
          *
          * @type BufferUsage
@@ -320,11 +333,13 @@ define([
         this._rs.depthTest.enabled = !this.showThroughEllipsoid;
 
         // Recompile shader when material changes
-        if (!this._material || (this._material !== this.material)) {
-
+        if (typeof this._material === 'undefined' ||
+            this._material !== this.material ||
+            this._affectedByLighting !== this.affectedByLighting) {
 
             this.material = (typeof this.material !== 'undefined') ? this.material : Material.fromType(context, Material.ColorType);
             this._material = this.material;
+            this._affectedByLighting = this.affectedByLighting;
 
             var fsSource =
                 '#line 0\n' +
@@ -333,6 +348,7 @@ define([
                 ShadersSensorVolume +
                 '#line 0\n' +
                 this._material.shaderSource +
+                (this._affectedByLighting ? '#define AFFECTED_BY_LIGHTING 1\n' : '') +
                 '#line 0\n' +
                 CustomSensorVolumeFS;