+++ Source/Shaders/ComplexConicSensorVolumeFS.glsl
@@ -38,7 +38,15 @@ vec4 getOuterColor(float sensorRadius, vec3 pointEC, vec3 normalEC)
     
     //Final
     vec3 positionToEyeEC = normalize(-v_positionEC);
-    return czm_lightValuePhong(czm_sunDirectionEC, positionToEyeEC, material);
+    
+    vec4 color; 
+    #ifdef AFFECTED_BY_LIGHTING    
+    color = czm_lightValuePhong(czm_sunDirectionEC, positionToEyeEC, material);
+    #else
+    color = vec4(material.diffuse, material.alpha);
+    #endif
+    
+    return color;        
 }
 
 vec4 getInnerColor(float sensorRadius, vec3 pointEC, vec3 normalEC)
@@ -48,7 +56,15 @@ vec4 getInnerColor(float sensorRadius, vec3 pointEC, vec3 normalEC)
     
     //Final
     vec3 positionToEyeEC = normalize(-v_positionEC);
-    return czm_lightValuePhong(czm_sunDirectionEC, positionToEyeEC, material);
+
+    vec4 color; 
+    #ifdef AFFECTED_BY_LIGHTING    
+    color = czm_lightValuePhong(czm_sunDirectionEC, positionToEyeEC, material);
+    #else
+    color = vec4(material.diffuse, material.alpha);
+    #endif
+    
+    return color;        
 }
 
 vec4 getCapColor(float sensorRadius, vec3 pointEC, vec3 normalEC)
@@ -58,7 +74,15 @@ vec4 getCapColor(float sensorRadius, vec3 pointEC, vec3 normalEC)
     
     //Final
     vec3 positionToEyeEC = normalize(-v_positionEC);
-    return czm_lightValuePhong(czm_sunDirectionEC, positionToEyeEC, material);
+
+    vec4 color; 
+    #ifdef AFFECTED_BY_LIGHTING    
+    color = czm_lightValuePhong(czm_sunDirectionEC, positionToEyeEC, material);
+    #else
+    color = vec4(material.diffuse, material.alpha);
+    #endif
+    
+    return color;        
 }
 
 vec4 getSilhouetteColor(float sensorRadius, vec3 pointEC, vec3 normalEC)
@@ -68,7 +92,15 @@ vec4 getSilhouetteColor(float sensorRadius, vec3 pointEC, vec3 normalEC)
     
     //Final
     vec3 positionToEyeEC = normalize(-v_positionEC);
-    return czm_lightValuePhong(czm_sunDirectionEC, positionToEyeEC, material);
+
+    vec4 color; 
+    #ifdef AFFECTED_BY_LIGHTING    
+    color = czm_lightValuePhong(czm_sunDirectionEC, positionToEyeEC, material);
+    #else
+    color = vec4(material.diffuse, material.alpha);
+    #endif
+    
+    return color;        
 }
 
 #endif