+++ Source/Shaders/CustomSensorVolumeFS.glsl
@@ -33,7 +33,15 @@ vec4 getColor(float sensorRadius, vec3 pointEC)
     materialInput.normalEC = normalEC;
     
     czm_material material = czm_getMaterial(materialInput);
-    return czm_lightValuePhong(czm_sunDirectionEC, positionToEyeEC, material);
+    
+    vec4 color; 
+    #ifdef AFFECTED_BY_LIGHTING    
+    color = czm_lightValuePhong(czm_sunDirectionEC, positionToEyeEC, material);
+    #else
+    color = vec4(material.diffuse, material.alpha);
+    #endif
+    
+    return color;        
 }
 
 #endif