+++ CHANGES.md
@@ -49,6 +49,8 @@ Beta Releases
 * Added support for CZML path visualization via the `DynamicPath` and `DynamicPathVisualizer` objects.  See the [CZML wiki](https://github.com/AnalyticalGraphicsInc/cesium/wiki/CZML-Guide) for more details.
 * Added support for [WEBGL_depth_texture](http://www.khronos.org/registry/webgl/extensions/WEBGL_depth_texture/).  See `Framebuffer.setDepthTexture`.
 * Added `CesiumMath.isPowerOfTwo`.
+* Added `affectedByLighting` to `ComplexConicSensorVolume`, `CustomSensorVolume`, and `RectangularPyramidSensorVolume` to turn lighting on/off for these objects.
+* CZML `Polygon`, `Cone`, and `Pyramid` objects are no longer affected by lighting.
 * Added `czm_viewRotation` and `czm_viewInverseRotation` automatic GLSL uniforms.
 
 ### b7 - 08/01/2012